removed last trig functions assignments added comments authorship details removed from storage.py - trig function aliases - unnecessary imports commented out pcm_temp()
