+++ CombiCSP/solar_tower.py
@@ -1,3 +1,9 @@
+# -*- coding: utf-8 -*- 
+"""
+    @Author: N. Papadakis
+    @Date: 2022/07/02
+    @Credit: original functions from G. Arnaoutakis
+"""
 import numpy as np
 import pandas as pd
 from CombiCSP import OutputContainer, CtoK
