+++ CombiCSP/solar_trough.py
@@ -1,6 +1,12 @@
+# -*- coding: utf-8 -*- 
+"""
+    @Author: N. Papadakis
+    @Date: 2022/07/02
+    @Credit: original functions from G. Arnaoutakis
+"""
+#%%
 import numpy as np
 import pandas as pd
-
 
 from CombiCSP import OutputContainer, CtoK
 import CombiCSP.SolarGeometry as sgh
